added command line table generation ability: KEYS: -table %diagram_textfile_path% %iterationCount% In the textfile should be only 2d diagram column heights split by space.
