+++ NDimYoung/Main.cpp
@@ -1,99 +1,11 @@
 #include "Tableaux2D.h"
 #include "TableSet.h"
 #include "YoungNDim_SimpleNode.h"
+#include "CommonMethods.h"
 
 using namespace std;
 
-#pragma region Common_Methods
 
-void ExportGraphLevel(vector<YoungNDim_SimpleNode*> level, ostream& stream)
-{
-	for (int i = 0; i < level.size(); i++)
-	{
-		level[i]->node->Print(stream);
-		stream << '|' << (*level[i]->dim) << '\n';
-	}
-}
-
-void ExportGraph(vector<vector<YoungNDim_SimpleNode*>> graph, ostream& stream)
-{
-	for (int i = 0; i < graph.size(); i++)
-	{
-		stream << "Level: " << i + 1 << "; diagrams count: " << graph[i].size() << '\n';
-		ExportGraphLevel(graph[i], stream);
-	}
-}
-
-//create graph with diagrams & their dimensions only 
-vector<vector<YoungNDim_SimpleNode*>> CreateGraph(YoungNDim_SimpleNode* rootDiagram, int levelCount, string exportOnFlyFileName = "\0")
-{
-	//result
-	vector<vector<YoungNDim_SimpleNode*>> levels(levelCount);
-	//add root diagram
-	levels[0] = { rootDiagram };
-	if (exportOnFlyFileName[0] != '\0') {
-		ofstream os(exportOnFlyFileName, ios_base::app);
-		os << "Level: 1; diagrams count: 1 \n";
-		ExportGraphLevel(levels[0], os);
-		os.close();
-	}
-	int streamNum = 4;
-	//iterate & create new levels based on previous ones
-	for (int i = 1; i < levelCount; i++)
-	{
-		//foreach diagram in previous level
-		for (int j = 0; j < levels[i - 1].size(); j++) {
-			//create children pool of selected
-			vector<YoungNDim_SimpleNode*> children = levels[i - 1][j]->Propagate();
-			//we can delete previously computed diagrams to save memory
-			if (exportOnFlyFileName[0] != '\0')
-				delete levels[i - 1][j];
-			//check new children (if they are already in the main pool)
-			for (int k = 0; k < levels[i].size(); k++)
-				for (int l = 0; l < children.size(); l++)
-					if (levels[i][k]->Equals(children[l]))
-					{
-						levels[i][k]->Absorb(children[l]);	//big pool absorbs a piece of smaller one
-						children.erase(children.begin() + l);	//remove absorbed diagram from check-pool
-						l--;	//step back (we destroyed last index)
-					}
-			//add new, unique children
-			levels[i].insert(levels[i].end(), children.begin(), children.end());
-			//go to next ancestor
-		}
-		levels[i - 1].clear();
-		//this level won't be expanded
-		levels[i].shrink_to_fit();
-		//export on fly
-		if (exportOnFlyFileName[0] != '\0') {
-			ofstream os(exportOnFlyFileName, ios_base::app);
-			os << "\n" << "Level: " << i + 1 << "; diagrams count: " << levels[i].size() << '\n';
-			ExportGraphLevel(levels[i], os);
-			os.close();
-		}
-		cout << "levels computed: " << i + 1 << '\n';
-		//all ancestors checked, go to next generation
-	}
-	return levels;
-}
-
-vector<string> split(const string& str, const string& delim)
-{
-	vector<string> tokens;
-	size_t prev = 0, pos = 0;
-	do
-	{
-		pos = str.find(delim, prev);
-		if (pos == string::npos)
-			pos = str.length();
-		string token = str.substr(prev, pos - prev);
-		//if (!token.empty()) //allow empty entries
-		tokens.push_back(token);
-		prev = pos + delim.length();
-	} while (pos < str.length() && prev < str.length());
-	return tokens;
-}
-#pragma endregion
 
 int main(int argc, char* argv[])
 {
@@ -115,23 +27,15 @@ int main(int argc, char* argv[])
 		printf("\n");
 	}*/
 	srand(0);
+
 	if (argc == 1)
 	{
 		int res = 0;
-		cout << "Enter 2D tableaux generating mode? [1 = yes, 0 = no]" << endl;
+		cout << "Enter 2D tableaux generating mode or N-dim graph builder? [1 = 2d tabl, 0 = graph mode]" << endl;
 		cin >> res;
 		if (res)
 		{
-			string dg;
-			Young2D diagram(0);
-			cout << "Enter column heights, split by [SPACE]:" << endl;
-			char c;
-			while (cin.get(c) && c != '\n');
-			getline(cin, dg);
-			vector<string> dgNms = split(dg, " ");
-			diagram.columns = vector<mainType>(dgNms.size());
-			for (int i = 0; i < dgNms.size(); i++)
-				diagram.columns[i] = stoi(dgNms[i]);
+			Young2D diagram(cin);
 			TableSet ts(diagram);
 			cout << "Enter generation iterations" << endl;
 			int genCnt = 0;
@@ -142,7 +46,7 @@ int main(int argc, char* argv[])
 			os.close();
 		}
 	}
-	bool infiniteGen = false;
+	bool infiniteGen = false, genTables = false;
 	int x = 0;
 	int graphDim = 4, lvls;
 	IYoung* startDiag;
@@ -156,6 +60,10 @@ int main(int argc, char* argv[])
 
 	for (; x < argc && !infiniteGen; x++)
 		infiniteGen = (strcmp("-all", argv[x]) == 0);
+	if(!infiniteGen)
+		for (x = 0; x < argc && !genTables; x++)
+			genTables = (strcmp("-tables", argv[x]) == 0);
+
 	//2nd key stands for dimension of the graph diagrams
 	if (infiniteGen) {
 		graphDim = atoi(argv[x]);
@@ -168,6 +76,19 @@ int main(int argc, char* argv[])
 			startDiag = new YoungNDim(1, graphDim);
 		CreateGraph(new YoungNDim_SimpleNode(startDiag), 999, fileName);
 		return 0;
+	}
+	
+	if (genTables)
+	{
+		ifstream str = ifstream(argv[x]);
+		Young2D diagram(str);
+		TableSet ts(diagram);
+		cout << "Enter generation iterations" << endl;
+		int genCnt = stoi(argv[x+1]);
+		ts.GenerateTables(genCnt);
+		ofstream os("Tableaux.txt");
+		ts.ExportTables(os);
+		os.close();
 	}
 	//ask dimensions count
 	cout << "Welcome to Young N-dimensonal graph builder!" << endl;
