+++ NDimYoung/Tableaux.txt
@@ -1,0 +1,16 @@
+0 67
+1 68
+2 53
+3 69
+4 58
+5 55
+6 76
+7 60
+8 54
+9 62
+10 69
+11 68
+12 50
+13 69
+14 68
+15 54
